(#191) - Split out temp views in a separate file
