+++ lib/index.js
@@ -24,6 +24,7 @@ modes.minimumForPouchDB = [
   'routes/revs-diff',
   'routes/root',
   'routes/session-stub',
+  'routes/temp-views',
   'routes/view-cleanup',
   'routes/views'
 ];
@@ -213,6 +214,7 @@ module.exports = function (startPouchDB, opts) {
     'routes/security',
     'routes/view-cleanup',
     'routes/views',
+    'routes/temp-views',
     'routes/ddoc-info',
     'routes/show',
     'routes/list',
